requirements.txt
transformer
numpy
pandas
matplotlib
tensorflow
streamlit

